Added maven test profiles
