+++ components/camel-minio/src/test/java/org/apache/camel/component/minio/integration/localTests/MinioComponentIntegrationTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.minio.integration.testContainers;
+package org.apache.camel.component.minio.integration.localTests;
 
 import org.apache.camel.EndpointInject;
 import org.apache.camel.Exchange;
