+++ components/camel-minio/src/test/java/org/apache/camel/component/minio/integration/localTests/MinioConsumerIntegrationTest.java
@@ -14,10 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.minio.integration;
-
-import java.io.IOException;
-import java.util.Properties;
+package org.apache.camel.component.minio.integration.localTests;
 
 import io.minio.MinioClient;
 import org.apache.camel.BindToRegistry;
@@ -25,20 +22,15 @@ import org.apache.camel.EndpointInject;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.minio.MinioConstants;
-import org.apache.camel.component.minio.MinioOperations;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.apache.camel.test.junit5.CamelTestSupport;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
-@Disabled("Must be manually tested. Goto https://play.min.io")
-public class MinioDeleteBucketOperationIntegrationTest extends CamelTestSupport {
-    final Properties properties = MinioTestUtils.loadMinioPropertiesFile();
+public class MinioConsumerIntegrationTest extends MinioTestContainerSupport {
 
     @BindToRegistry("minioClient")
-    MinioClient minioClient = MinioClient.builder()
-            .endpoint(properties.getProperty("endpoint"))
-            .credentials(properties.getProperty("accessKey"), properties.getProperty("secretKey"))
+    MinioClient client = MinioClient.builder()
+            .endpoint("http://" + CONTAINER.getHost(), CONTAINER.getMappedPort(BROKER_PORT), false)
+            .credentials(ACCESS_KEY, SECRET_KEY)
             .build();
 
     @EndpointInject
@@ -47,19 +39,23 @@ public class MinioDeleteBucketOperationIntegrationTest extends CamelTestSupport 
     @EndpointInject("mock:result")
     private MockEndpoint result;
 
-    public MinioDeleteBucketOperationIntegrationTest() throws IOException {
-    }
-
     @Test
     public void sendIn() throws Exception {
-        result.expectedMessageCount(1);
+        result.expectedMessageCount(3);
 
-        template.send("direct:listBuckets",
-                exchange -> exchange.getIn().setHeader(MinioConstants.MINIO_OPERATION, MinioOperations.listBuckets));
+        template.send("direct:putObject", exchange -> {
+            exchange.getIn().setHeader(MinioConstants.OBJECT_NAME, "test1.txt");
+            exchange.getIn().setBody("Test1");
+        });
 
-        template.send("direct:deleteBucket", exchange -> {
-            exchange.getIn().setHeader(MinioConstants.BUCKET_NAME, "mycamel2");
-            exchange.getIn().setHeader(MinioConstants.MINIO_OPERATION, MinioOperations.deleteBucket);
+        template.send("direct:putObject", exchange -> {
+            exchange.getIn().setHeader(MinioConstants.OBJECT_NAME, "test2.txt");
+            exchange.getIn().setBody("Test2");
+        });
+
+        template.send("direct:putObject", exchange -> {
+            exchange.getIn().setHeader(MinioConstants.OBJECT_NAME, "test3.txt");
+            exchange.getIn().setBody("Test3");
         });
 
         assertMockEndpointsSatisfied();
@@ -70,11 +66,11 @@ public class MinioDeleteBucketOperationIntegrationTest extends CamelTestSupport 
         return new RouteBuilder() {
             @Override
             public void configure() {
-                String minioEndpoint = "minio://mycamel2?autoCreateBucket=true";
+                String minioEndpoint = "minio://mycamel?autoCreateBucket=true";
 
-                from("direct:listBuckets").to(minioEndpoint);
-
-                from("direct:deleteBucket").to(minioEndpoint).to("mock:result");
+                from("direct:putObject").startupOrder(1).to(minioEndpoint);
+                from("minio://mycamel?moveAfterRead=true&destinationBucketName=camel-kafka-connector&autoCreateBucket=true")
+                        .startupOrder(2).to("mock:result");
 
             }
         };
