+++ components/camel-minio/src/test/java/org/apache/camel/component/minio/integration/localTests/MinioListObjectsOperationIntegrationTest.java
@@ -14,11 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.minio.integration;
+package org.apache.camel.component.minio.integration.localTests;
 
 import java.io.IOException;
 import java.util.Iterator;
-import java.util.Properties;
 
 import com.google.common.collect.Iterators;
 import io.minio.MinioClient;
@@ -33,20 +32,16 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.minio.MinioConstants;
 import org.apache.camel.component.minio.MinioOperations;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.apache.camel.test.junit5.CamelTestSupport;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@Disabled("Must be manually tested. Goto https://play.min.io")
-public class MinioListObjectsOperationIntegrationTest extends CamelTestSupport {
-    final Properties properties = MinioTestUtils.loadMinioPropertiesFile();
+public class MinioListObjectsOperationIntegrationTest extends MinioTestContainerSupport {
 
     @BindToRegistry("minioClient")
-    MinioClient minioClient = MinioClient.builder()
-            .endpoint(properties.getProperty("endpoint"))
-            .credentials(properties.getProperty("accessKey"), properties.getProperty("secretKey"))
+    MinioClient client = MinioClient.builder()
+            .endpoint("http://" + CONTAINER.getHost(), CONTAINER.getMappedPort(BROKER_PORT), false)
+            .credentials(ACCESS_KEY, SECRET_KEY)
             .build();
 
     @EndpointInject
