+++ components/camel-minio/src/test/java/org/apache/camel/component/minio/integration/remoteTests/MinioCopyObjectOperationIntegrationTest.java
@@ -14,7 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.minio.integration.testContainers;
+package org.apache.camel.component.minio.integration.remoteTests;
+
+import java.io.IOException;
+import java.util.Properties;
 
 import io.minio.MinioClient;
 import org.apache.camel.BindToRegistry;
@@ -23,15 +26,18 @@ import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.minio.MinioConstants;
 import org.apache.camel.component.minio.MinioOperations;
+import org.apache.camel.component.minio.integration.MinioTestUtils;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.test.junit5.CamelTestSupport;
 import org.junit.jupiter.api.Test;
 
-public class MinioDeleteBucketOperationIntegrationTest extends MinioTestContainerSupport {
+public class MinioCopyObjectOperationIntegrationTest extends CamelTestSupport {
+    final Properties properties = MinioTestUtils.loadMinioPropertiesFile();
 
     @BindToRegistry("minioClient")
-    MinioClient client = MinioClient.builder()
-            .endpoint("http://" + CONTAINER.getHost(), CONTAINER.getMappedPort(BROKER_PORT), false)
-            .credentials(ACCESS_KEY, SECRET_KEY)
+    MinioClient minioClient = MinioClient.builder()
+            .endpoint(properties.getProperty("endpoint"))
+            .credentials(properties.getProperty("accessKey"), properties.getProperty("secretKey"))
             .build();
 
     @EndpointInject
@@ -40,16 +46,23 @@ public class MinioDeleteBucketOperationIntegrationTest extends MinioTestContaine
     @EndpointInject("mock:result")
     private MockEndpoint result;
 
+    public MinioCopyObjectOperationIntegrationTest() throws IOException {
+    }
+
     @Test
     public void sendIn() throws Exception {
         result.expectedMessageCount(1);
 
-        template.send("direct:listBuckets",
-                exchange -> exchange.getIn().setHeader(MinioConstants.MINIO_OPERATION, MinioOperations.listBuckets));
+        template.send("direct:putObject", exchange -> {
+            exchange.getIn().setHeader(MinioConstants.OBJECT_NAME, "test.txt");
+            exchange.getIn().setBody("Test");
+        });
 
-        template.send("direct:deleteBucket", exchange -> {
-            exchange.getIn().setHeader(MinioConstants.BUCKET_NAME, "mycamel2");
-            exchange.getIn().setHeader(MinioConstants.MINIO_OPERATION, MinioOperations.deleteBucket);
+        template.send("direct:copyObject", exchange -> {
+            exchange.getIn().setHeader(MinioConstants.OBJECT_NAME, "test.txt");
+            exchange.getIn().setHeader(MinioConstants.DESTINATION_OBJECT_NAME, "test1.txt");
+            exchange.getIn().setHeader(MinioConstants.DESTINATION_BUCKET_NAME, "mycamel1");
+            exchange.getIn().setHeader(MinioConstants.MINIO_OPERATION, MinioOperations.copyObject);
         });
 
         assertMockEndpointsSatisfied();
@@ -60,11 +73,11 @@ public class MinioDeleteBucketOperationIntegrationTest extends MinioTestContaine
         return new RouteBuilder() {
             @Override
             public void configure() {
-                String minioEndpoint = "minio://mycamel2?autoCreateBucket=true";
+                String minioEndpoint = "minio://mycamel?autoCreateBucket=false";
 
-                from("direct:listBuckets").to(minioEndpoint);
+                from("direct:putObject").to(minioEndpoint);
 
-                from("direct:deleteBucket").to(minioEndpoint).to("mock:result");
+                from("direct:copyObject").to(minioEndpoint).to("mock:result");
 
             }
         };
