+++ components/camel-minio/src/test/java/org/apache/camel/component/minio/integration/remoteTests/MinioDeleteBucketOperationIntegrationTest.java
@@ -14,8 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.minio.integration;
+package org.apache.camel.component.minio.integration.remoteTests;
 
+import java.io.IOException;
 import java.util.Properties;
 
 import io.minio.MinioClient;
@@ -24,20 +25,19 @@ import org.apache.camel.EndpointInject;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.minio.MinioConstants;
+import org.apache.camel.component.minio.MinioOperations;
+import org.apache.camel.component.minio.integration.MinioTestUtils;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.camel.test.junit5.CamelTestSupport;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
-@Disabled("Goto https://play.min.io and search for 'mycamel' and 'camel-kafka-connector'. If bucket(s) does not exist, set 'autoCreateBucket=true' in route(s)")
-public class MinioConsumerIntegrationTest extends CamelTestSupport {
+public class MinioDeleteBucketOperationIntegrationTest extends CamelTestSupport {
     final Properties properties = MinioTestUtils.loadMinioPropertiesFile();
 
     @BindToRegistry("minioClient")
-    MinioClient client = MinioClient.builder()
+    MinioClient minioClient = MinioClient.builder()
             .endpoint(properties.getProperty("endpoint"))
             .credentials(properties.getProperty("accessKey"), properties.getProperty("secretKey"))
-            .region(properties.getProperty("region"))
             .build();
 
     @EndpointInject
@@ -46,26 +46,19 @@ public class MinioConsumerIntegrationTest extends CamelTestSupport {
     @EndpointInject("mock:result")
     private MockEndpoint result;
 
-    public MinioConsumerIntegrationTest() throws Exception {
+    public MinioDeleteBucketOperationIntegrationTest() throws IOException {
     }
 
     @Test
     public void sendIn() throws Exception {
-        result.expectedMessageCount(3);
+        result.expectedMessageCount(1);
 
-        template.send("direct:putObject", exchange -> {
-            exchange.getIn().setHeader(MinioConstants.OBJECT_NAME, "test1.txt");
-            exchange.getIn().setBody("Test1");
-        });
+        template.send("direct:listBuckets",
+                exchange -> exchange.getIn().setHeader(MinioConstants.MINIO_OPERATION, MinioOperations.listBuckets));
 
-        template.send("direct:putObject", exchange -> {
-            exchange.getIn().setHeader(MinioConstants.OBJECT_NAME, "test2.txt");
-            exchange.getIn().setBody("Test2");
-        });
-
-        template.send("direct:putObject", exchange -> {
-            exchange.getIn().setHeader(MinioConstants.OBJECT_NAME, "test3.txt");
-            exchange.getIn().setBody("Test3");
+        template.send("direct:deleteBucket", exchange -> {
+            exchange.getIn().setHeader(MinioConstants.BUCKET_NAME, "mycamel2");
+            exchange.getIn().setHeader(MinioConstants.MINIO_OPERATION, MinioOperations.deleteBucket);
         });
 
         assertMockEndpointsSatisfied();
@@ -76,11 +69,11 @@ public class MinioConsumerIntegrationTest extends CamelTestSupport {
         return new RouteBuilder() {
             @Override
             public void configure() {
-                String minioEndpoint = "minio://mycamel?autoCreateBucket=false";
+                String minioEndpoint = "minio://mycamel2?autoCreateBucket=true";
 
-                from("direct:putObject").startupOrder(1).to(minioEndpoint);
-                from("minio://mycamel?moveAfterRead=true&destinationBucketName=camel-kafka-connector&autoCreateBucket=false")
-                        .startupOrder(2).to("mock:result");
+                from("direct:listBuckets").to(minioEndpoint);
+
+                from("direct:deleteBucket").to(minioEndpoint).to("mock:result");
 
             }
         };
