+++ components/camel-minio/src/test/java/org/apache/camel/component/minio/integration/remoteTests/MinioListObjectsOperationIntegrationTest.java
@@ -14,10 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.minio.integration.testContainers;
+package org.apache.camel.component.minio.integration.remoteTests;
 
 import java.io.IOException;
 import java.util.Iterator;
+import java.util.Properties;
 
 import com.google.common.collect.Iterators;
 import io.minio.MinioClient;
@@ -31,17 +32,20 @@ import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.minio.MinioConstants;
 import org.apache.camel.component.minio.MinioOperations;
+import org.apache.camel.component.minio.integration.MinioTestUtils;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.test.junit5.CamelTestSupport;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class MinioListObjectsOperationIntegrationTest extends MinioTestContainerSupport {
+public class MinioListObjectsOperationIntegrationTest extends CamelTestSupport {
+    final Properties properties = MinioTestUtils.loadMinioPropertiesFile();
 
     @BindToRegistry("minioClient")
-    MinioClient client = MinioClient.builder()
-            .endpoint("http://" + CONTAINER.getHost(), CONTAINER.getMappedPort(BROKER_PORT), false)
-            .credentials(ACCESS_KEY, SECRET_KEY)
+    MinioClient minioClient = MinioClient.builder()
+            .endpoint(properties.getProperty("endpoint"))
+            .credentials(properties.getProperty("accessKey"), properties.getProperty("secretKey"))
             .build();
 
     @EndpointInject
